requirements.txt
asgiref @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_edg6nzjrx0/croots/recipe/asgiref_1661287549316/work
backports.zoneinfo @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4ampz1_6cv/croots/recipe/backports.zoneinfo_1661287537252/work
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet @ file:///opt/concourse/worker/volumes/live/c798b2ee-88b1-4341-6830-161a92c2399e/volume/chardet_1607706832595/work
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_066uop7vqa/croots/recipe/django_1661292540858/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1677341444254/work
django-templated-mail==1.1.1
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.4.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==2.6.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
python-dateutil==2.8.2
python3-openid==3.2.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_204sy4e5t2/croot/sqlparse_1668793448973/work
stripe @ file:///Users/ktietz/demo/mc3/conda-bld/stripe_1629463074029/work
uritemplate==4.1.1
urllib3==1.26.15
